requirements.txt
asgiref>=3.2.10
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django>=3.2.24
django-crispy-forms==2.1
django-storages==1.14.3
django-taggit==3.0.0
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django
social-auth-core
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
